Continue porting to clojure.java-time
I was almost able to complete this process as I have learned how to
use the library well, but there is a field missing. I was going to
open an issue against the library about that, but I found that it had
been fixed already, and as soon as commit access is granted to the
people who want to build a new release, I will be able to update to
that and finish.